+++ projects/ng-core-ui/src/lib/main.ts
@@ -1,6 +1,7 @@
 import { EnvironmentProviders, makeEnvironmentProviders, InjectionToken } from '@angular/core';
 import { withInterceptors, provideHttpClient } from '@angular/common/http';
 import { MatPaginatorIntl } from '@angular/material/paginator';
+import { MAT_DATE_LOCALE, provideNativeDateAdapter } from '@angular/material/core';
 import { contextInterceptor } from './system/context-interceptor';
 import { loadingInterceptor } from './system/loading.interceptor';
 import { GpaPaginatorIntl } from './system/gpa-paginator-intl';
@@ -19,5 +20,7 @@ export function provideGPAUICore(options?: GpaCoreOptions): EnvironmentProviders
     ...(options?.environmentUrl ? [{ provide: LIB_ENVIRONMENT_URL, useValue: options.environmentUrl }] : []),
     provideHttpClient(withInterceptors([contextInterceptor, loadingInterceptor])),
     { provide: MatPaginatorIntl, useClass: GpaPaginatorIntl },
+    { provide: MAT_DATE_LOCALE, useValue: 'it' },
+    provideNativeDateAdapter(),
   ]);
 }